+++ quilt-sync/ui/src/components.rs
@@ -1,6 +1,7 @@
 pub mod buttons;
 pub mod ignore_popup;
 pub mod layout;
+pub mod popover;
 pub mod set_remote_popup;
 pub mod spinner;
 pub mod update_checker;
@@ -8,6 +9,7 @@ pub mod workflow_select;
 
 pub use ignore_popup::{IgnorePopup, IgnorePopupData, UnignorePopup, UnignorePopupData};
 pub use layout::{Layout, Notification, ToolbarActions};
+pub use popover::with_popover;
 pub use set_remote_popup::{SetRemotePopup, SetRemotePopupData};
 pub use spinner::Spinner;
 pub use update_checker::UpdateChecker;